adding ml_sampler_with_histogram_weights package
Summary:
Motivation behind adding this package: As discussed in this post https://fburl.com/fr47zwto, there is a need for slightly more flexibility with the ml_sampler. Adding a new package for histogram weights will be helpful in constructing more context specific pdfs.

Opting not to modify the currently ml_smapler.py because other pipelines currently rely on the 4 column input.

Differential Revision: D8917427

fbshipit-source-id: 5f7b42ce4800603adcf484fdefcbee196853424c